+++ crates/cli/src/commands/link.rs
@@ -11,7 +11,7 @@ pub async fn link(directory: Option<PathBuf>) -> Result<()> {
 
     if config.token.is_none() {
         return Err(anyhow!(
-            "You are not logged in. Please log in with `lagon login`",
+            "You are not logged in. Please log in with `lagoss login`",
         ));
     }