+++ crates/runtime_http/src/headers.rs
@@ -3,5 +3,5 @@ pub const X_FORWARDED_PROTO: &str = "x-forwarded-proto";
 pub const X_FORWARDED_HOST: &str = "x-forwarded-host";
 pub const X_REAL_IP: &str = "x-real-ip";
 
-pub const X_LAGON_REGION: &str = "x-lagon-region";
-pub const X_LAGON_ID: &str = "x-lagon-id";
+pub const X_LAGOSS_REGION: &str = "x-lagoss-region";
+pub const X_LAGOSS_ID: &str = "x-lagoss-id";