+++ crates/wpt-runner/src/main.rs
@@ -1,8 +1,8 @@
 use console::style;
 use hyper::{http::Request, Body};
-use lagon_runtime::{options::RuntimeOptions, Runtime};
-use lagon_runtime_http::RunResult;
-use lagon_runtime_isolate::{options::IsolateOptions, Isolate, IsolateEvent, IsolateRequest};
+use lagoss_runtime::{options::RuntimeOptions, Runtime};
+use lagoss_runtime_http::RunResult;
+use lagoss_runtime_isolate::{options::IsolateOptions, Isolate, IsolateEvent, IsolateRequest};
 use std::{
     env, fs,
     path::{Path, PathBuf},